refactor: reorganize service handler layer
