+++ custom_components/thessla_green_modbus/services_handlers.py
@@ -2,6 +2,7 @@
 
 from .services_handler_deps import ServiceHandlerDeps
 from .services_handlers_data import register_data_services
+from .services_handlers_logging import register_logging_services
 from .services_handlers_maintenance import register_maintenance_services
 from .services_handlers_mode import register_mode_services
 from .services_handlers_parameters import register_parameter_services
@@ -10,6 +11,7 @@ from .services_handlers_schedule import register_schedule_services
 __all__ = [
     "ServiceHandlerDeps",
     "register_data_services",
+    "register_logging_services",
     "register_maintenance_services",
     "register_mode_services",
     "register_parameter_services",
