+++ custom_components/thessla_green_modbus/services_helpers.py
@@ -5,12 +5,8 @@ from __future__ import annotations
 from collections.abc import Callable
 from typing import Any, cast
 
-import voluptuous as vol
 from homeassistant.core import HomeAssistant, ServiceCall
 from homeassistant.helpers.service import async_extract_entity_ids
-
-from .const import DOMAIN
-from .modbus_exceptions import ConnectionException, ModbusException
 
 
 def extract_entity_ids(hass: HomeAssistant, call: ServiceCall) -> set[str]:
@@ -35,50 +31,6 @@ def iter_target_coordinators(
     return targets
 
 
-def validate_bypass_temperature_range(data: dict[str, Any]) -> dict[str, Any]:
-    """Validate bypass temperature range independently from voluptuous internals."""
-    temperature = data.get("min_outdoor_temperature")
-    if temperature is not None and not (-20.0 <= float(temperature) <= 40.0):
-        invalid_exc = getattr(vol, "Invalid", ValueError)
-        raise invalid_exc(f"min_outdoor_temperature ({temperature}) must be in range -20.0..40.0")
-    return data
-
-
-def validate_gwc_temperature_range(data: dict[str, Any]) -> dict[str, Any]:
-    """Reject configurations where min_air_temperature >= max_air_temperature."""
-    tmin = data.get("min_air_temperature")
-    tmax = data.get("max_air_temperature")
-    if tmin is not None and tmax is not None and tmin >= tmax:
-        invalid_exc = getattr(vol, "Invalid", ValueError)
-        raise invalid_exc(
-            f"min_air_temperature ({tmin}) must be strictly less than max_air_temperature ({tmax})"
-        )
-    return data
-
-
-def normalize_option(value: str) -> str:
-    """Convert translation keys to internal option values."""
-    if value and value.startswith(f"{DOMAIN}."):
-        value = value.split(".", 1)[1]
-    prefixes = [
-        "special_mode_",
-        "day_",
-        "period_",
-        "bypass_mode_",
-        "gwc_mode_",
-        "filter_type_",
-        "reset_type_",
-        "modbus_port_",
-        "modbus_baud_rate_",
-        "modbus_parity_",
-        "modbus_stop_bits_",
-    ]
-    for prefix in prefixes:
-        if value.startswith(prefix):
-            return value[len(prefix) :]
-    return value
-
-
 def clamp_airflow_rate(coordinator: Any, airflow_rate: int) -> int:
     """Clamp airflow_rate to the coordinator's reported min/max percentages."""
     data = getattr(coordinator, "data", {}) or {}
@@ -97,19 +49,3 @@ def clamp_airflow_rate(coordinator: Any, airflow_rate: int) -> int:
     if max_val < min_val:
         max_val = min_val
     return max(min_val, min(max_val, int(airflow_rate)))
-
-
-async def write_register(
-    coordinator: Any,
-    register: str,
-    value: Any,
-    entity_id: str,
-    action: str,
-    logger: Any,
-) -> bool:
-    """Write to a register with error handling."""
-    try:
-        return bool(await coordinator.async_write_register(register, value, refresh=False))
-    except (ModbusException, ConnectionException) as err:
-        logger.error("Failed to %s for %s: %s", action, entity_id, err)
-        return False
